modularlize `gix-blame/lib.rs`
